Isolate is_call_possible_via_execution_policy in its own header
