+++ agency/detail/bulk_functions/bulk_then.hpp
@@ -1,13 +1,19 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
-#include <agency/detail/tuple.hpp>
-#include <agency/detail/type_list.hpp>
-#include <agency/future.hpp>
 #include <agency/detail/type_traits.hpp>
+#include <agency/detail/type_list.hpp>
 #include <agency/detail/bulk_functions/decay_parameter.hpp>
 #include <agency/detail/bulk_functions/unpack_shared_parameters_from_executor_and_invoke.hpp>
 #include <agency/detail/bulk_functions/execute_agent_functor.hpp>
+#include <agency/detail/bulk_functions/scope_result.hpp>
+#include <agency/detail/bulk_functions/single_result.hpp>
+#include <agency/detail/bulk_functions/result_factory.hpp>
+#include <agency/detail/bulk_functions/bind_agent_local_parameters.hpp>
+#include <agency/detail/execution_policy_traits.hpp>
+#include <agency/execution_agent.hpp>
+#include <agency/executor_traits.hpp>
+#include <utility>
 
 namespace agency
 {
